FLIP-136 [feat] PullToRefresh 리팩토링
1. 상태 변화(화면 회전 등) 시 무한 새로고침 혹은 새로고침 중단 이슈 해결
